+++ app/src/main/kotlin/org/monsing/chat/MessageDto.kt
@@ -1,4 +1,4 @@
-package org.monsing.application
+package org.monsing.chat
 
 data class MessageDto(
     val chatId: String,
