+++ build.gradle
@@ -33,15 +33,23 @@ subprojects {
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'org.springframework.boot:spring-boot-starter-validation'
         implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+        implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
         implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
         implementation 'org.jetbrains.kotlin:kotlin-reflect'
+        implementation 'org.springframework.boot:spring-boot-starter-websocket'
         runtimeOnly 'com.h2database:h2'
         runtimeOnly 'com.mysql:mysql-connector-j'
+
+        //test
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
         testImplementation 'io.kotest:kotest-runner-junit5-jvm:5.9.1'
         testImplementation 'io.kotest:kotest-assertions-core-jvm:5.9.1'
+        testImplementation 'de.flapdoodle.embed:de.flapdoodle.embed.mongo.spring30x:4.6.2'
+        testImplementation 'com.github.codemonstur:embedded-redis:1.4.3'
+        testImplementation 'io.mockk:mockk:1.13.13'
 
         //그라파나, 프로메테우스, 액츄에이터
         implementation 'org.springframework.boot:spring-boot-starter-actuator:3.3.5'
