+++ chat/src/main/kotlin/org/monsing/api/WebSocketHandler.kt
@@ -1,6 +1,6 @@
 package org.monsing.api
 
-import org.monsing.application.ChatService
+import org.monsing.chat.ChatService
 import org.springframework.stereotype.Component
 import org.springframework.web.socket.CloseStatus
 import org.springframework.web.socket.WebSocketMessage
