+++ chat/src/main/resources/application.yml
@@ -3,15 +3,4 @@ spring:
     active: local
     include:
       - app
-
-  data:
-    mongodb:
-      username:
-      password:
-      uri: mongodb://localhost:27017/chat
-
-    redis:
-      host: localhost
-      port: 6379
-      username:
-      password:
+      - domain
