+++ domain/src/main/kotlin/org/monsing/chat/Chat.kt
@@ -1,4 +1,4 @@
-package org.monsing.domain
+package org.monsing.chat
 
 import jakarta.persistence.Id
 import java.util.UUID
