+++ domain/src/main/kotlin/org/monsing/config/RedisProperties.kt
