+++ domain/src/main/resources/application-domain.yml
@@ -1,0 +1,16 @@
+spring:
+  config:
+    activate:
+      on-profile: local
+
+  data:
+    mongodb:
+      username:
+      password:
+      uri: mongodb://localhost:27017/chat
+
+    redis:
+      host: localhost
+      port: 6379
+      username:
+      password:
